Exercise definitions refactor (#283)
Simply the way that exercises are being defined and logic is being reused, so it's much easier to follow the logic. There is more simplification that can be done, particularly with isolating pure logic so that high level code is shorter and easier to understand. The PR descriptions has more suggestions. 

* Exercise definitions refactor: intervals and basic infra

* Migrate more exercises (progress)

* Restore the defaults to be separate

* Add anroid/.idea to gitignore

* triad inversion exercise

* Chord types

* Remove "Compose exercise"

* Notes with chords exercise

* Chord in key exercise

* notesInKeyExercise

* common chord progressions

* Chords in real songs exercise

* Use roman analysis exercise in "common chords"

* Avoid nesting when using plugable exercise utils

* Introduce "settingsConfig" to group settings related config

* romanAnalysis settingsConfig

* Pr fixes

* Simplify answerList

* Stabilize adaptive mode

* Remove unnecessary code